+++ src/app/about/page.tsx
@@ -1,10 +1,12 @@
 import PageHeaderAbout from "@/components/about/PageHeaderAbout";
+import MyStorySection from "@/components/about/MyStorySection";
 
 
 export default function AboutPage() {
     return (
         <>
             <PageHeaderAbout />
+            <MyStorySection/>
         </>
     )
 }
